microservices/markdown_analysis_service/requirements.txt
# Markdown Analysis Service
# Install with: pip install -r requirements.txt

# FastAPI server
fastapi>=0.110.0
uvicorn>=0.23.0

# OpenAI API client
openai>=1.0.0

# Data validation
pydantic>=2.0.0

# Environment variable loading
python-dotenv>=1.0.0
